Correct the output by converting it into a `Long` as the multiplication of two large integers can give a `Long` result. (E.g., 1000000 * 90000)

+++ src/level10/maxPairProduct.kt
@@ -54,5 +54,5 @@ fun main() {
 
     // Printing the product of the two largest numbers available
 //    println("The product of the two largest number is: ${firstLargestNumber * secondLargestNumber}")
-    println(firstLargestNumber * secondLargestNumber)
+    println((firstLargestNumber * secondLargestNumber).toLong())
 }
